\name{oz-states}
\title{Plot Specific Australian States}
\alias{nsw}
\alias{nt}
\alias{qld}
\alias{sa}
\alias{tas}
\alias{vic}
\alias{wa}
\description{
  Plot the named Australian state along with its coastline.
}
\usage{
nsw(\dots)
nt(\dots)
qld(\dots)
sa(\dots)
tas(\dots)
vic(\dots)
wa(\dots)
}
\arguments{
  \item{\dots}{arguments to be passed on to \code{\link{oz}}.}
}
\keyword{hplot} 
